chore: changed file structure
