+++ App.js
@@ -1,85 +1,85 @@
 
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-import { Pedometer } from 'expo-sensors';
+// import React from 'react';
+// import { StyleSheet, Text, View } from 'react-native';
+// import { Pedometer } from 'expo-sensors';
 
-export default class App extends React.Component {
-  state = {
-    isPedometerAvailable: 'checking',
-    pastStepCount: 0,
-    currentStepCount: 0,
-  };
+// export default class App extends React.Component {
+//   state = {
+//     isPedometerAvailable: 'checking',
+//     pastStepCount: 0,
+//     currentStepCount: 0,
+//   };
 
-  componentDidMount() {
-    this._subscribe();
-  }
+//   componentDidMount() {
+//     this._subscribe();
+//   }
 
-  componentWillUnmount() {
-    this._unsubscribe();
-  }
+//   componentWillUnmount() {
+//     this._unsubscribe();
+//   }
 
-  _subscribe = () => {
-    this._subscription = Pedometer.watchStepCount(result => {
-      this.setState({
-        currentStepCount: result.steps,
-      });
-    });
+//   _subscribe = () => {
+//     this._subscription = Pedometer.watchStepCount(result => {
+//       this.setState({
+//         currentStepCount: result.steps,
+//       });
+//     });
 
-    Pedometer.isAvailableAsync().then(
-      result => {
-        this.setState({
-          isPedometerAvailable: String(result),
-        });
-      },
-      error => {
-        this.setState({
-          isPedometerAvailable: 'Could not get isPedometerAvailable: ' + error,
-        });
-      }
-    );
+//     Pedometer.isAvailableAsync().then(
+//       result => {
+//         this.setState({
+//           isPedometerAvailable: String(result),
+//         });
+//       },
+//       error => {
+//         this.setState({
+//           isPedometerAvailable: 'Could not get isPedometerAvailable: ' + error,
+//         });
+//       }
+//     );
 
-    const end = new Date();
-    const start = new Date();
-    start.setDate(end.getDate() - 1);
-    Pedometer.getStepCountAsync(start, end).then(
-      result => {
-        this.setState({ pastStepCount: result.steps });
-      },
-      error => {
-        this.setState({
-          pastStepCount: 'Could not get stepCount: ' + error,
-        });
-      }
-    );
-  };
+//     const end = new Date();
+//     const start = new Date();
+//     start.setDate(end.getDate() - 1);
+//     Pedometer.getStepCountAsync(start, end).then(
+//       result => {
+//         this.setState({ pastStepCount: result.steps });
+//       },
+//       error => {
+//         this.setState({
+//           pastStepCount: 'Could not get stepCount: ' + error,
+//         });
+//       }
+//     );
+//   };
 
-  _unsubscribe = () => {
-    this._subscription && this._subscription.remove();
-    this._subscription = null;
-  };
+//   _unsubscribe = () => {
+//     this._subscription && this._subscription.remove();
+//     this._subscription = null;
+//   };
 
-  render() {
-    return (
-      <View style={styles.container}>
-        <Text>Pedometer.isAvailableAsync(): {this.state.isPedometerAvailable}</Text>
-        <Text>Steps taken in the last 24 hours: {this.state.pastStepCount}</Text>
-        <Text>Walk! And watch this go up: {this.state.currentStepCount}</Text>
-        <Text>Walk 5 steps to get to the next phase:</Text>
-    {this.state.currentStepCount < 5 ? <Text>Start</Text> : <Text>Next Stage</Text>}
-    {this.state.currentStepCount < 5 ? <View style={{height: 100, width: 100, backgroundColor: 'rgb(0,255,0)'}} /> : <View style={{height: 100, width: 100, backgroundColor: 'rgb(255,0,0)'}}/> }
-      </View>
-    );
-  }
-}
+//   render() {
+//     return (
+//       <View style={styles.container}>
+//         <Text>Pedometer.isAvailableAsync(): {this.state.isPedometerAvailable}</Text>
+//         <Text>Steps taken in the last 24 hours: {this.state.pastStepCount}</Text>
+//         <Text>Walk! And watch this go up: {this.state.currentStepCount}</Text>
+//         <Text>Walk 5 steps to get to the next phase:</Text>
+//     {this.state.currentStepCount < 5 ? <Text>Start</Text> : <Text>Next Stage</Text>}
+//     {this.state.currentStepCount < 5 ? <View style={{height: 100, width: 100, backgroundColor: 'rgb(0,255,0)'}} /> : <View style={{height: 100, width: 100, backgroundColor: 'rgb(255,0,0)'}}/> }
+//       </View>
+//     );
+//   }
+// }
 
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    marginTop: 15,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+// const styles = StyleSheet.create({
+//   container: {
+//     flex: 1,
+//     marginTop: 15,
+//     alignItems: 'center',
+//     justifyContent: 'center',
+//   },
+// });
